### 2022-07-15 #550 @markean
Submission - melt: Multiple Empirical Likelihood Tests
Submitting Author Name: Eunseop Kim
Submitting Author Github Handle: @markean
Repository:  https://github.com/markean/melt
Version submitted: 1.6.0.9000
Submission type: Stats
Badge grade: silver
Version accepted: TBD
---

```
Package: melt
Title: Multiple Empirical Likelihood Tests
Version: 1.6.0.9000
    person("Eunseop", "Kim", , "kim.7302@osu.edu", role = c("aut", "cre")),
    person("Steven", "MacEachern", role = c("ctb", "ths")),
    person("Mario", "Peruggia", role = c("ctb", "ths"))
  )
Description: Performs multiple empirical likelihood tests for linear and
    generalized linear models. The core computational routines are
    implemented using the 'Eigen' C++ library and 'RcppEigen' interface,
    with OpenMP for parallel computation. Details of multiple testing
    procedures are given in Kim, MacEachern, and Peruggia (2021)
    <arxiv:2112.09206>.
    graphics,
    Rcpp,
    stats
    covr,
    microbenchmark,
    withr
    BH,
    dqrng,
    Rcpp,
    RcppEigen
    knitr
NeedsCompilation: yes
Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd",
    "srr::srr_stats_roclet"))

```

- [x] A pre-submission inquiry has been approved in [issue #549](#<issue_num>)

The package attempts the first implementation of the nested bilevel optimization approach within R to compute constrained empirical likelihood. The inner layer Newton-Raphson method for empirical likelihood is written in C++, enabling faster computation than other routines written in R.

Yes.

Silver.

1) Compliance with a good number of standards beyond those identified as minimally necessary.
2) Have a demonstrated generality of usage beyond one single envisioned use case.
In my opinion, the generality comes from the applicability of empirical likelihood methods to (generalized) linear models and the ability to test linear hypotheses of choice.

The package is on CRAN.
