### 2022-08-05 #550 @markean
Hi @pchausse. Thank you so much for reviewing our package. Please let me know if you have any question. A draft version of pdf vignette in `vignettes` folder might be of help.
