### 2022-08-09 #550 @markean
Hi @awstringer1! Thank you for reviewing the package. Please let me know if you have any question. A draft version of pdf vignette in `vignettes` folder might be of help.

@pchausse and @awstringer1: By an external request, a newer version of the package will be released to CRAN in August. It will not interfere with the review process, but I will tag a development version after the release. Then the reviewer comments will be reflected to the package in a subsequent release.
