### 2022-08-10 #550 @awstringer1
I am not sure if this is something the authors did, or something the `rOpenSci` template does, but the code line references in the "srr report" do not seem to match up with the locations of the tags in the `roxygen` documentation. They are a few lines off. This is fine mostly but sometimes I cannot find the relevant tag, e.g. for "Standards in on line#306 of file [R/AllGenerics.R](https://github.com/markean/melt/blob/master/R/AllGenerics.R#L306):" Can the authors comment? Thank you.
