### 2022-08-10 #550 @mpadge
@awstringer1 That's because the repo has been updated since the report linked to above was generated. As stated in the automated checks, you can - *and in this case should* - generate a local, and up-to-date, version of the report by running `srr::srr_report(path)`, where `path` is where your local clone of `melt` is. By default that will open the full report in your web browser, in which all reported line numbers should match current state of repo.
