### 2022-08-10 #550 @markean
@awstringer1 The `path` should be your local clone of the package, not the installed path. I can generate the report with `srr::srr_report(path)`. I guess that is what @mpadge meant above.
