### 2022-08-13 #550 @markean
@pchausse and @awstringer1: v1.7.0 has been released to CRAN. No changes were made with respect to the `srr` tags. The code on the repository will be updated after reviewers’ comments. Thank you.
