### 2022-08-17 #550 @awstringer1
Here is my review, I hope it is helpful!

---

### Compliance with Standards

- [x] This package complies with a sufficient number of standards for a silver badge
- [x] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: (Please specify)

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

For packages aiming for silver or gold badges:

- [x] This package extends beyond minimal compliance with standards in the following ways: (please describe)

The package seems to have been developed for future general use in mind. The authors have provided an interface that is clearly intended to be used beyond the examples considered in their associated preprint. A large number of the standards described in the `rOpenSci` review materials seem to have been considered by the authors when designing their software.

#### Comments

Overall, the authors appear to have convincingly motivated that the package meets the `rOpenSci` standards. Many statements in the authors' `@srrstats` documentation tags contain incomplete references to the package documentation, which precludes a more thorough review. See below for a non-exhaustive list of examples. I have the following comments about the authors' justifications:

- _"@srrstats {RE1.4} Some of the core distributional assumptions are documented."_: this is potentially confusing to users. Either all of the regularity conditions should be clearly documented, or a link to a specific section in a paper where these conditions are clearly documented should be given. Although arguably common in the literature, the phrase "under some regularity conditions" is incomplete and not meaningful, and I suggest its removal.

- _"@srrstats {G2.1, G2.1a} Assertions on types of inputs are clarified throughout the package documentation."_: it would be helpful to the reviewers to give specific references to the package documentation. There are a number of similar "throughout the package documentation" statements that render it challenging to assess compliance with standards, beyond a general sense of such.

- Possible typo: "na.cation()" in `el_glm.R`, line 64.

- _"@srrstats {RE1.0, RE1.1} Formula interface is used to the `formula` argument, and how it is converted to a matrix input is documented as well."_: a reference to the function, and section of documentation where this is located would better facilitate review.

- _"@srrstats {RE4.11} `logLik()` method extracts the empirical log-likelihood value which can be interpreted as a measure of goodness-of-fit."_: more detail could be provided about how this is to be used as a measure of goodness of fit, perhaps a link to where this is discussed in the package documentation or the preprint?

- _"@srrstatsNA {G1.5} The package does not have any associated publication."_: is it appropriate to cite the preprint here (this might be a question for the Editor)?

- _"@srrstatsNA {RE2.3} It is beyond the scope of estimating functions that the package considers. There seems to be no simple way to pass this information to the internal routines. Users can use `el_eval()` to manually construct estimating functions with the offset included, but any further method not applicable to the output of `el_eval()`."_: this appears to be in response to standard _RE2.3_ which states: _"RE2.3 Where applicable, Regression Software should enable data to be centred (for example, through converting to zero-mean equivalent values; or to z-scores) or offset (for example, to zero-intercept equivalent values) via additional parameters, with the effects of any such parameters clearly documented and tested."_. The present package uses `lm` and `glm`, as well as a standard formula interface, all of which facilitates centering and offsets, so it is not clear why these tasks are out of scope for the present package.

---

### General Review

I do not see such a statement. I think this statement could be helpful: a short paragraph describing why empirical likelihood methods are useful in linear and generalized linear models could be informative to the user, since (a) empirical likelihood is a somewhat "specialized" topic and (b) there are other methods available for doing hypothesis tests in linear and generalized linear models, and such methods also rely on limiting Chi-square distributional arguments. It is at least not trivially clear why empirical likelihood is useful and why we'd want a package for it, so I think being very up front about this could better motivate the package.

- [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING

I see no such statement.

- [ ] The documentation is sufficient to enable general use of the package beyond one specific use case

The documentation is almost there, but I am going to leave this unchecked for now. I think adding more detailed "Details" in _all_ core functions, even if this involves a lot of copy and pasting, is a good idea, see (https://r-pkgs.org/man.html)[https://r-pkgs.org/man.html]. For example, `?el_lm` contains very detailed documentation, but `?el_mean` does not. It is helpful to users to have everything on one page, even if the material is repeated across many functions' documentation.

#### Algorithms

The core algorithms are implemented in `C++`. I will do my best to provide appropriate comments here, but I am not an expert on `C++`.

- `EL.cpp` line 103: is there justification for this step-halving mechanism? It is only applied once within the loop, so how can it ensure that the function value is increased? Would a `while` loop be more appropriate, or perhaps use of an algorithm that guarantees the desired property?

- Related: how often does the Newton step here not increase the function value in practice? If often, perhaps Newton's method is unsuitable here, and a Trust Region procedure could be considered?

- `EL.cpp` line 181: is explicitly creating a projection matrix necessary? The combination of this being a dense matrix, and relying on explicitly calling `inverse` (which may have performance and stability implications), leads me to question whether operations involving this matrix could instead be coded more efficiently. For example, if we have P = I - X(X X^T)^-1 X^T, and you want to compute a matrix/vector product Pv, you could write Pv = v - Xb where (XX^T)b = X^Tv, and b is obtained by solving this system. This avoids the need to ever compute and store an explicit inverse.

- What is the purpose of the PsuedoLog class? Commenting this code (lines 347--394) could help facilitate a more thorough review. This looks like a generic mathematical operation-- is there no pre-existing implementation to rely on?

The testing appears thorough, I have no comments.

#### Visualisation (where appropriate)

The visualizations provided seem to serve their stated purpose. No comments.

#### Package Design

**External**: the package is designed for the specific task of computing empirical likelihood ratio tests and intervals for the parameters of linear and generalized linear models. The package is set up to take input similar to `lm` and `glm` and then call the associated `.fit` function from core `R`. Some suggestions for broadening scope:

- Could methods for objects of class `lm` and `glm` (and so on) be provided, so that the package provides its output to users who have already fit their models using these alternative methods, and hence compare them?
- The output of `el_lm` is an object of class `LM`, a new class from the `melt` package. Could this output be made to inherit from the core `R` `lm` class, to use the associated functionality with which most `R` users are familiar? Or related, could the package provide a straightforward manner by which to fit an `lm` object from a fitted `LM` object? Again, thinking about a user who wishes to compare the two. Similar comments for `glm`.

**Internal**: see comments about algorithms above. In general the use of `C++` for the numerical analysis seems appropriate, notwithstanding the specific comments above.

---
