### 2022-08-26 #550 @markean
Hi @awstringer1, thank you for your review and feedback. We have made many changes to the main branch based on your comments. The package is in a better shape than before. Please find the responses below.

## Response to @awstringer1
#### Comments
> - _"@srrstats {RE1.4} Some of the core distributional assumptions are documented."_: this is potentially confusing to users. Either all of the regularity conditions should be clearly documented, or a link to a specific section in a paper where these conditions are clearly documented should be given. Although arguably common in the literature, the phrase "under some regularity conditions" is incomplete and not meaningful, and I suggest its removal.

The phrase is removed. Interested users are directed to the references for more details.

> - _"@srrstats {G2.1, G2.1a} Assertions on types of inputs are clarified throughout the package documentation."_: it would be helpful to the reviewers to give specific references to the package documentation. There are a number of similar "throughout the package documentation" statements that render it challenging to assess compliance with standards, beyond a general sense of such.

It is documented in `@param` sections of el_mean.R, where the relevant `srr` tags are attached as well. These tags can be also located in any other user-facing functions of the package. All the user-facing functions state the required types of inputs explicitly.

> - Possible typo: "na.cation()" in `el_glm.R`, line 64.

Corrected to `na.action()`.

> - _"@srrstats {RE1.0, RE1.1} Formula interface is used to the `formula` argument, and how it is converted to a matrix input is documented as well."_: a reference to the function, and section of documentation where this is located would better facilitate review.

It is documented in `@details` section of el_lm.R, where the relevant `srr` tags are attached as well. Explanation on how the `formula` input relates to the `model.frame` and `model.matrix` objects is not provided since there is nothing special about the usage of `formula` in the package. Interested users can refer to the documentation of `model.frame()` or `model.matrix()` of the `stats` package.

> - _"@srrstats {RE4.11} `logLik()` method extracts the empirical log-likelihood value which can be interpreted as a measure of goodness-of-fit."_: more detail could be provided about how this is to be used as a measure of goodness of fit, perhaps a link to where this is discussed in the package documentation or the preprint?

The `srr` tag is moved to log.R. `logL()` and `logLR()` extract empirical log-likelihood (ratio) at a specified parameter value.
The formulation of empirical likelihood problem can be viewed as allocating probabilities to each observation while minimizing a goodness-of-fit criterion (Cressie–Read discrepancy) subject to constraints. Hence, the computed empirical log-likelihood (ratio) values can also be interpreted as a goodness-of-fit measure. @references sections are added to the package documentation in AllGenerics.R (lines 404--406, 441--443).

> - _"@srrstatsNA {RE2.3} It is beyond the scope of estimating functions that the package considers. There seems to be no simple way to pass this information to the internal routines. Users can use `el_eval()` to manually construct estimating functions with the offset included, but any further method not applicable to the output of `el_eval()`.": this appears to be in response to standard RE2.3 which states: "RE2.3 Where applicable, Regression Software should enable data to be centred (for example, through converting to zero-mean equivalent values; or to z-scores) or offset (for example, to zero-intercept equivalent values) via additional parameters, with the effects of any such parameters clearly documented and tested."_. The present package uses `lm` and `glm`, as well as a standard formula interface, all of which facilitates centering and offsets, so it is not clear why these tasks are out of scope for the present package.

`el_lm()` and `el_glm()` now accepts an optional `offset` argument. With offsets, the estimating functions were over-identified in the previous package design, and over-identified estimating equations are beyond the scope of the package. All the related C++ functions have been rewritten to accommodate offsets. The functionality with `offset` has been tested internally, and it works with other methods of the package. However, it is still in an experimental stage, and further changes could be made in a future release.
<br/><br/>

#### Documentation
> - [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README

> I do not see such a statement. I think this statement could be helpful: a short paragraph describing why empirical likelihood methods are useful in linear and generalized linear models could be informative to the user, since (a) empirical likelihood is a somewhat "specialized" topic and (b) there are other methods available for doing hypothesis tests in linear and generalized linear models, and such methods also rely on limiting Chi-square distributional arguments. It is at least not trivially clear why empirical likelihood is useful and why we'd want a package for it, so I think being very up front about this could better motivate the package.

`Overview` section of README file is expanded to address the comment.

> - [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING

> I see no such statement.

CONTRIBUTING is in the .github folder since it is a GitHub specific file.

> The documentation is sufficient to enable general use of the package beyond one specific use case

> The documentation is almost there, but I am going to leave this unchecked for now. I think adding more detailed "Details" in all core functions, even if this involves a lot of copy and pasting, is a good idea, see (https://r-pkgs.org/man.html)[https://r-pkgs.org/man.html]. For example, `?el_lm` contains very detailed documentation, but `?el_mean` does not. It is helpful to users to have everything on one page, even if the material is repeated across many functions' documentation.

`@details` section is added in el_mean.R.
<br/><br/>

#### Algorithms
> `EL.cpp` line 103: is there justification for this step-halving mechanism? It is only applied once within the loop, so how can it ensure that the function value is increased? Would a `while` loop be more appropriate, or perhaps use of an algorithm that guarantees the desired property?

> Related: how often does the Newton step here not increase the function value in practice? If often, perhaps Newton's method is unsuitable here, and a Trust Region procedure could be considered?

Much of the code for the computation is based on Owen (2001, chapter 12).

Here, `set_el()` performs different tasks depending on whether the convex hull constraint is satisfied or not. When the convex hull constraint is satisfied, the objective function is strictly concave without any constraint. Thus, the Newton–Raphson method converges quadratically to the unique solution. Step-halving is a safe-guard against possible numerical instability.

When the convex hull constraint is violated, the `PsuedoLog` class (explained below) increases the function value until it reaches a set threshold, or until the algorithm hits the maximum iteration number. Since the optimization problem is infeasible in this case, a while loop for the step-halving often runs indefinitely. The underlying principle is that we do not check the convex hull constraint explicitly for the `EL` class. Indirectly inspecting the constraint from the output is much faster and more efficient.

Currently, the package only considers infinitely differentiable estimating functions, and evaluating empirical likelihood itself with Newton–Raphson method is not a challenging task. Other slower but safer methods such as the Quasi-Newton or the Levenberg-Marquardt can be considered in the future when needed.

> `EL.cpp` line 181: is explicitly creating a projection matrix necessary? The combination of this being a dense matrix, and relying on explicitly calling `inverse` (which may have performance and stability implications), leads me to question whether operations involving this matrix could instead be coded more efficiently. For example, if we have P = I - X(X X^T)^-1 X^T, and you want to compute a matrix/vector product Pv, you could write Pv = v - Xb where (XX^T)b = X^Tv, and b is obtained by solving this system. This avoids the need to ever compute and store an explicit inverse.

New `proj()` in helpers.cpp (lines 31--35) performs the same projection operation. It does not compute an explicit inverse.

Below is a benchmark result of the same projection operation with a random 5 by 50 LHS matrix. The version without the inverse is much faster, and the gap will grow as the dimension of the matrix increases.
```
Unit: microseconds
        expr   min    lq     mean  median     uq      max neval cld
     inverse 9.491 9.906 11.25626 10.7715 12.307   45.142 10000   b
w/o inverse 3.073 3.293  4.49282  3.5470  4.102 7289.781 10000  a
```

> What is the purpose of the PsuedoLog class? Commenting this code (lines 347--394) could help facilitate a more thorough review. This looks like a generic mathematical operation-- is there no pre-existing implementation to rely on?

A brief comment is added to EL.h (lines 129--137).

Numerical computation of empirical log-likelihood ratio statistics heavily relies on a pseudo-log function. It makes the associated optimization problem easier to solve without affecting the solution when the convex hull constraint is satisfied. When the condition is violated, the pseudo-log function also prevents the statistic from being undefined (Inf) so that the algorithm can fail gracefully.
The  `PsuedoLog` class has relevant member variables and static member functions needed for the computation. We have not found any pre-existing routines for these specific tasks.
<br/><br/>

#### Package Design
> Could methods for objects of class `lm` and `glm` (and so on) be provided, so that the package provides its output to users who have already fit their models using these alternative methods, and hence compare them?

Providing empirical likelihood methods to existing `lm` or `glm` classes directly would require dedicated functions like `summary_el()` or `confint_el()`. Otherwise, they would mask existing ones like `summary()` or `confint()`, which would not be desirable. Although `el_lm()` and `el_glm()` purposefully mimic the syntax and output of `lm()` and `glm()`, these functions perform significance tests, validate inputs, and prepare objects for subsequent methods provided by the package. Since the options provided by `el_lm()` and `el_glm()` are different from `lm()` and `glm()`, currently we are not considering extending methods to existing S3 classes.

> The output of `el_lm` is an object of class `LM`, a new class from the `melt` package. Could this output be made to inherit from the core `R` `lm` class, to use the associated functionality with which most `R` users are familiar? Or related, could the package provide a straightforward manner by which to fit an `lm` object from a fitted `LM` object? Again, thinking about a user who wishes to compare the two. Similar comments for `glm`.

The idea of inheriting from `lm` class was considered in an early stage of development, but it was abandoned due to the following reasons:

- Many generics and methods that apply to `lm` are not applicable to `LM` by the characteristics of empirical likelihood methods.
These include, for example, `cooks.distance()`, `dfbeta()`, `extractAIC()`, `anova()`, `influence()`, `rstudent()`, and `vcov()`.

- For linear and generalized linear models, some methods would overlap with `lm` if `LM` inherits from `lm`, meaning that they would produce the same output. Examples include `coef()`, `resid()`, and `predict()`. Although these could be added, we strive to keep the package focused on testing and inference, not on estimation or prediction.

- Too many generics and methods are applicable to `lm`. Over 30 methods are applicable to `lm` from the `base` and `stats` packages.

We plan to add some of the methods to `LM` directly in a future release.

> Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

We are grateful for your insightful comments and suggestions on the package. Your name and role were added to the updated DESCRIPTION file. We tagged you in NEWS file for the update on projection operation. We would also like to include @pchausse when we receive comments from both reviewers.
