### 2022-09-08 #550 @markean
Hi @pchausse, thank you for your review and feedback. Your comments have greatly improved the documentation. Some notable updates on the code include: a new `logProb()` function is added to extract the `logp` slot, `plot()` is directly applicable to `EL` class, `elt()` allows a symbolic description for specifying a hypothesis, `print()` shows the tested hypothesis when applied to `ELT` class. The repository currently has v1.7.0.9600 of the package. The CRAN version will be updated after the review process. Please find the responses below.

## Response to @pchausse
### Compliance with Standards
> **G1.6 Software should include code necessary to compare performance claims with alternative implementations in other R packages.**: Other packages compute empirical likelihood tests and confidence intervals. We  would like to see how the melt package compares with the others. For example, my momentfit (or gmm) package produces the same type of confidence intervals or confidence areas for linear models and the mean. The package emplik now includes regression models.

Since we do not make performance claims compared to other packages, please check the benchmark result below for reference only. We constructed EL confidence intervals for the mean with 100 observations using the melt, momentfit, and emplik packages. Default settings were used for each package.
```
library(microbenchmark)
library(melt)
library(emplik)
library(momentfit)
set.seed(134095)
x <- rnorm(100)
fit <- el_mean(x, 0)
out <- microbenchmark(
  ci_melt = confint(fit),
  ci_momentfit = confint(x, gelType = "EL"),
  ci_emplik = findUL(fun = function(theta, x) {
    el.test(x, mu = theta)
  }, MLE = mean(x), x = x),
  times = 1000
)
print(out, unit = "relative")
```
```
Unit: relative
         expr      min        lq      mean    median        uq       max neval cld
      ci_melt  1.00000  1.000000  1.000000  1.000000  1.000000  1.000000  1000 a
 ci_momentfit  6.18093  5.738152  5.672995  5.549923  5.304345  5.693226  1000  b
    ci_emplik 33.49088 29.969031 31.332941 30.790802 31.826105 69.604799  1000   c
```
<br/><br/>

### Comments
> **"@srrstats {RE1.4} Some of the core distributional assumptions are documented."**: The list of regularity conditions is not that long. We do see some distribution assumptions in the `EL` documentation where the EL method is presented for general estimating equations $E[g(X, \theta)] = 0$, but it would be important to add at least the assumption on the variance of $g(X, \theta)$. Users using only `el_mean` or `el_sd` should know what distribution assumptions are required. For these two functions, the detail section is empty. At least, we should see what the estimating equations are and then be given a link to the EL documentation for more details.

The updated documentation of `el_mean()` and `el_sd()` state in `Details` section the optimization problems with the estimating functions. Assumptions on the variance are stated as well. The `See Also` section gives a link to the `EL` documentation. Links to relevant documentation have been added to all the other user-facing exported functions as well.

> **"@srrstats {RE4.11} `logLik()` method extracts the empirical log-likelihood value which can be interpreted as a measure of goodness-of-fit"**: It is really not clear how to interpret the value. Can we have a more detailed explanation?

`logLik()` only gives $-n\log n$ with $p_i = 1/n$ for all $i$. We have expanded the `Details` section of `logLik()` (line 436 of AllGenerics.R) with more details. `logLik()` was originally designed to offer other functionality, but its role has been superseded by `logL()`. The `srr` tag has also been moved to logL-methods.R.

> **"@srrstats {RE6.0} `plot` method is available for an `ELD` object that is returned by the `eld()` method."**: The authors may consider adding some diagnostic-type plot methods for `EL` (or any other) class as we have for `lm` objects. It is great to have a plot for `eld()` to detect outliers, but it would be even better I think if the plot was applied directly to `EL` class. For example, a Cook Distance exists separately for `lm`, but we can plot it directly from `lm` objects by selecting `which=4`:

`eld()` is now applicable to `EL` objects directly (line 35 of plot-methods.R). It implicitly calls `eld()` since the package currently provides no other diagnostic plot.  More options like `which` can be added in a future version.
<br/><br/>

### General Review
#### Documentation
> I think we should see a more detailed presentation of why one would need to use the `melt` package to perform hypothesis tests or to construct confidence intervals. For example, when the empirical likelihood starts to produce confidence intervals with better coverage than `lm` does for linear regression? In the example section of `el_lm`, we see an application with iid and normally distributed variables. In this case, even when the same size is very small, a least square confidence interval has an exact coverage while EL does not. So why should we use the `melt` package?

Please check the updated README file. We have added more details in the `Overview` section. The new `Main functions` section introduces some available functions with short descriptions. More examples have also been added in the `Usage` section. The examples of `el_lm()` now rely on a new internal dataset `thiamethoxam`.

##### The `optim` slot
> What is `par`? What is `lambda`? Since the model is just-identified, the latter should be a vector of zeroes when there are no constraints. Are they the estimates when we restrict all coefficients to be zero? (except for the intercept). Can that be added to the documentation?

The updated `Details` sections of `LM`, `GLM`, and `QGLM` contain more information on the `optim` slot and its components. The following slots are all related to the overall test: `optim`, `logp`, `logl`, `loglr`, `statistic`, `df`, and `pval`.

> In the `el_lm` and `el_glm` (maybe others?) it is written in the details that the test for the null $H_0 : \theta_1 = \theta_2 = \cdots = \theta_{p - 1} = 0$ is in `optim` and the test parameter by parameter is in `sigTests`, but there is no test result in `optim`. It only includes the parameters, the lambdas, the number of iterations and the convergence code. Also, why is it important to return the implied probabilities (the `logp` slot) and the lambdas for the joint test and not for the individual ones?

Besides `optim`, the overall test result needs to involve `logp`, `logl`, `loglr`, `statistic`, `df`, and `pval`. Since they are all derived from `optim`, we stated that the overall test result is in `optim`. In order to avoid confusion, we have removed the original sentence "The test results are returned as `optim` and `sigTests`, respectively." from `el_lm()` and `el_glm()`. The updated documentation of `el_lm()` and `el_glm()` now focuses on what it does rather than specific slots of its return value which are not explained.

We can include the full results(`optim`, `logp`, `logl`, `loglr`, `statistic`, `df`, and `pval`) for each significance test, but it will make the `sigTests` slot have nested structure with possibly large size. Currently, `sigTest()` applied to `SummaryLM` returns the estimates, the test statistics, and the $p$-values. `elt()` can be if the user wants more information on a specific test.

Additionally, we have added a new accessor method `logProb()` that extracts `logp`.

##### The other EL slots
> The slots are documented in EL, but it is not easy to get it for other objects. For example, in the `el_glm` documentation, the Value is "An object of class GLM". If we click in it, we only see that it inherits from LM. If we click on it, we see some slots, but not all. It inherits from CEL, so we click on it and only the `optim` slot is defined. We finally get the description of all slots by clicking on EL. It is not an easy process from someone like me who only work in terminals. It requires to call `help()` several times. It would be helpful if the slots were defined for all objects that inherits from EL.

> Also, it is not always clear what the slots are. In EL, we see "A numeric vector of the log probabilities obtained from empirical likelihood" for the `logp` slot. But which probabilities? I figured out for `el_lm` that it is for the restricted model when all parameters are 0 except the intercept. It should be specified. It is the same for `logl`, `loglr`, etc. This is something we see across the different help files. It is another reason the author should define all slots for all functions that return an EL object.

The updated documentation (in AllClasses.R) lists all slots for all classes with some class-specific details.

##### The Empirical Likelihood method
> I would expect a little more details about the EL method. It is well presented in the EL documentation for general estimating equations $g(X, \theta)$, but I think it would be helpful for the method to also be explained in `el_lm`, `el_glm`, `el_sd`, `el_mean` and even `el_eval`. The EL method does not need to be repeated, but at least we should see the specific estimating equation used for each model and a link to EL for the details. For example, it is not clear to me which estimating equation is used for the standard deviation or for generalized linear models.

Each model-building function (including `el_eval()`) now states estimating equations and the associated EL problem in the `Details` section.

`el_eval()` is a function that can be used to compute EL when the user has a specific estimating function not covered by the package. For a just-identified estimating function $g$ with a specified parameter $\theta$, the user needs to prepare the input matrix `g` computed with data. `el_eval()` only solves the optimization problem and returns the result. Since the main functions, such as `elt()` and `elmt()`, cannot be used with an arbitrary estimating equation, the return value is a `list` with the same components of `EL`.

##### Other comments
> As a suggestion, the package can be greatly improved by adding more general nonlinear regression models. In `el_glm` is a nonlinear model, so it should bot be to difficult to include nonlinear regression like in `nls` or even more general cases. With a method to estimate general nonlinear models, it becomes straightforward to test nonlinear restrictions, if they are limited to restrictions that can be written as $H_0: \theta_i = f(\theta_{-i})$. if not too difficult to implement because it only implies replacing $\theta_i$ by $f(\theta_i)$ in the estimating equation function $g(X, \theta)$.

Although we agree that `nls` is probably the most natural way for an extension of the package, the development is currently focused on adding families and link functions for `el_glm()`. The next release version (v1.8.0) will be shipped with `identity` link option for `quasipoisson` family, which was added recently in C++ files. So it will take several more releases for `el_glm()` to provide most of the options in our roadmap.

Based on your comments above, `nls` methods could be implemented by improving `el_eval()` or making new functions such as `el_nls()` or `el_g()`. These functions should take a `formula` or `function` argument for estimating functions and an additional argument for gradient functions. Then, in principle, we could construct confidence intervals or perform hypothesis testing. The main concern would be performance. The package exclusively relies on the Eigen C++ library and OpenMP directives for parallel computing, and user-defined R functions have no simple thread-safe way of accessing them. We will revisit this issue when the development of `el_glm()` is finished.

#### Algorithms

##### The `confint` method
> It seems that all methods for the `EL`, `QGLM` and `SD` classes are
almost identical and can be reduced to a single method applied to a
union class. I can see some benefits from avoiding repeating the same
codes in several methods, especially if changes need to be applied to
them.
>
> Why the getMethodEL(object) is used to get the type of method for
the classes EL and `QGLM` and not for the class `SD`? Is it because
the getMethodEL method does not exist for this class? For the
latter, the type is explicitly added (see line 214 of
confint-methods.R)

`confint()` methods for `QGLM` and `SD` classes have been removed. One method for `EL` class can be applied to all other classes using a new internal generic `getEstimates()` (line 663 of AllGenerics.R). We have applied the same procedure to `confreg()` and `elmt()` methods and reduced the amount of duplicate code.

##### The `elt` method
> Would be nice to see the hypothesis being tested when the result is
printed (see for example the linearHypothesis function from the
car package).

`pirnt()` method now prints the tested hypothesis. A new internal function `describe_hypothesis()` in describe_hypothesis.R translates the hypothesis matrix into a character vector for printing. Additionally, `elt()` now accepts a character vector for the argument `lhs`, allowing a symbolic description of a hypothesis. The updated documentation of elt.R shows different ways to specify a hypothesis with new examples.

> Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

Please check the updated DESCRIPTION file. We have also added you and @awstringer1 in the `Acknowledgments` section of the pdf vignette in `vignettes/article.pdf`.
