### 2022-09-11 #550 @awstringer1
Thanks to @markean for the response. I was waiting until both reviews and responses were in, to avoid redundancy. All my major comments have been addressed.

My only minor comment is about the optimization. I don't think it is true that "_the objective function is strictly concave without any constraint. Thus, the Newton–Raphson method converges quadratically to the unique solution._"– see Nocedal and Wright's _Numerical Optimization_, Theorem 3.5, page 46. There are non-trivial initial conditions, namely that the starting value has to be "close enough" to the optimum, else it will diverge. If Newton with step-halving has worked so far, then that's good enough for me (and perhaps there is some additional theory here about why it works in this context that I do not know), but if it ever becomes a practical problem in the future, the authors could consider trust region optimization. See Chapter 4 in that same book, or the `trust` or `trustOptim` `R` packages. I have found this very useful in cases where Newton was not working satisfactorily.
